feature 0.1 :P

+++ src/main/resources/db/changelog/0.0.1/changes/0.1-start.sql
@@ -6,6 +6,6 @@ CREATE TABLE IF NOT EXISTS users.users
     username varchar,
     password varchar,
     email varchar,
-    balance int,
+    balance decimal,
     is_frozen boolean
 );
